requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
